backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
pyinstaller
pywebview
pandas
openpyxl
requests
